+++ src/InDesign/Command/DynamicLayoutBreakInterface.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * mds Agenturgruppe GmbH
+ *
+ * This source file is licensed under GNU General Public License version 3 (GPLv3).
+ *
+ * Full copyright and license information is available in
+ * LICENSE.md which is distributed with this source code.
+ *
+ * @copyright  Copyright (c) mds. Agenturgruppe GmbH (https://www.mds.eu)
+ */
+
+namespace Mds\PimPrint\CoreBundle\InDesign\Command;
+
+/**
+ * Interface DynamicLayoutBreakInterface
+ *
+ * @package Mds\PimPrint\CoreBundle\InDesign\Command
+ */
+interface DynamicLayoutBreakInterface
+{
+}